Fix crash when using emoji args in btn and btnp
