+++ source/emojiconversion.h
@@ -1,0 +1,3 @@
+#include <string>
+
+std::string convert_emojis(const std::string& lua);
